requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.0
nltk>=3.6.2
joblib>=1.0.1
requests>=2.25.1
python-dotenv>=0.19.0 